third_party/WebKit/LayoutTests/web-animations-api/animation-constructor.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<div id='e'></div>

<script>
var element = document.getElementById('e');

var keyframes = [{opacity: '1'}, {opacity: '0'}];
var timingObject = {duration: 2, iterations: 5};

test(function() {
    var animation = new Animation(element, keyframes, timingObject);
    assert_not_equals(animation, undefined);
    assert_equals(animation.constructor, Animation);
}, 'Calling new Animation() with a timing object input should create an animation.');

test(function() {
    var animation = new Animation(element, keyframes, 2);
    assert_not_equals(animation, undefined);
    assert_equals(animation.constructor, Animation);
}, 'Calling new Animation() with a duration input should create an animation.');

test(function() {
    var animation = new Animation(element, keyframes);
    assert_not_equals(animation, undefined);
    assert_equals(animation.constructor, Animation);
}, 'Calling new Animation() with no timing input should create an animation.');

test(function() {
    var animation = new Animation(null, keyframes);
    assert_not_equals(animation, undefined);
    assert_equals(animation.constructor, Animation);
}, 'Calling new Animation() with no target should create an animation.');
</script>
